+++ src/main/java/lotto/ui/ResultView.java
@@ -1,6 +1,6 @@
 package lotto.ui;
 
-import lotto.domain.Lotto;
+import lotto.domain.lotto.Lotto;
 import lotto.domain.prize.Prize;
 
 import java.util.List;
