+++ src/test/java/lotto/domain/lotto/LottoMachineTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoMachine;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
